Export a binary from so-acceptance module
This helps to clean up the command used for executing a test run in dependent modules. It also decouples from codeceptjs since this is no longer exposed to any implmentations.

This has the additional benefit that since it no longer depends on a particular relative install location for codeceptjs, the restriction for npm versions can be lifted.