Pass config file location as `--config` flag to codecept
This means that the root directory for the test run defaults to the current working directory - i.e. the project directory - instead of the `so-acceptance` install directory. This makes it easier to pass custom, per-project config because the paths in project config are relative to their own directory and not `so-acceptance`.