Remove dependency on hof-bootstrap
Loading all of hof-bootstrap into the dependency tree to get access to one small module was kinda nasty - especially when the major versions weren't in sync. A *good* way of solving this would be to abstract the module into a standalone npm thing, but that feels like overkill for something which is highly unlikely to change.