Refactor configuration file API
Codecept since 0.6.0 will no longer default to `process.cwd()` when running tests, and instead infers the test root from the config file location - which is not configurable.

This means that the config file in the project root directory *must* be passed to codecept, and so the extension of the common config in this module must be explicitly extended.